enhance hint data validation
